requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
django-ckeditor-filebrowser-filer==0.2.0b12
django-classy-tags==2.0.0
django-cms==3.8.0
django-crispy-forms==1.11.2
django-filebrowser==3.13.2
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-grappelli==2.14.3
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-tinymce==3.3.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-text-ckeditor==4.0.0
easy-thumbnails==2.7.1
filter-and-pagination==1.1.14
html5lib==1.1
idna==2.10
jsmin==2.2.2
mysql-client==0.0.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.2
protobuf==3.15.8
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.2
urllib3==1.26.4
webencodings==0.5.1
